feat: Add version completions for specified package
This commit adds lookups for tagged versions of packages. It adds
the caret (^) and tilde (~) qualifiers for major or minor versions
as options. Unfortunately, because of the way Panic triggers and
filters completions, I can't use the completion feature for
suggestions. I thought it would be easier for users to 'search' by
package tags (e.g., 'latest') instead of version numbers, since it's
easy to lose track of JavaScript version numbers.