fix: Use regex for package search instead of `context.text`
If searching for a package with a dash in it (fairly common on npm),
then Nova's `context.text` would only provide characters after the
dash. This commit uses a regex match instead to capture what is
likely the full search query. It also improves said regex for other
characters that might be in a package name.