refactor: Add optional range to version completion generation
In mocking the CompletionItem class, I discovered it has the 'range'
property for declaring what a completion should replace. Not sure
how I missed it before. Anyways, this property could come in handy
for replacing trailing quotes or commas after the cursor when
completing a package version. So, I'm adding it to the method and
the tests.