refactor: Switch to using snippets for insert text
Something about trying to helpfully reformat the JSON punctuation
and whitespace was causing serious issues. It seemed to work better
to use snippets for inserting text. This also helped wean the
extension from using ranges for completions, which another extension
author identified as a source of completions not showing. Another
benefit of snippets is that they let me auto-suggest the lastest
version immediately after the package is chosen. As you can see,
this was a fairly fundamental change so it touched about every
script file.