fix: Revert to using a range for the package name
Any npm package with kebab-case (of which there are many) would
not be properly inserted without providing a range that includes
any characters before the hyphen. Only characters after the last
hyphen would be replaced. So, packages do need to use ranges, so
I'm putting back the earlier code to generate a range for the
package name completion.