test: Add test for ensuring full package name is replaced by range
Not very TDD, but as I was replacing code previously written it's
maybe not so bad that I'm now testing the behavior I want. This
test ensures that kebab-cased package names are fully replaced
when users select a completion. Also adding checks in a previous
test that non-kebabed packages also have a proper range.